requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
openai>=1.40.0
anthropic>=0.34.0
streamlit>=1.38.0
pandas>=2.2.0
pydantic>=2.8.0
python-dotenv>=1.0.0
langsmith>=0.1.100